myapp/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
SQLAlchemy==2.0.34
httpx==0.27.2
pytest==8.3.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.6